Externalize inner classes from OAuth2ClientAutoConfiguration
